requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
scikit-surprise
streamlit